Refactor tf
