+++ lib/ros/be/infra/files/terraform/aws/eks-resources/outputs.tf
@@ -1,0 +1,4 @@
+
+output "alb_arn" {
+  value = data.external.alb_arn.result.LoadBalancerArn
+}
