+++ lib/ros/be/infra/files/terraform/aws/globalaccelerator/outputs.tf
@@ -4,8 +4,3 @@ output "globalaccelerator_ips" {
     aws_globalaccelerator_accelerator.this.*.ip_sets.0.ip_addresses,
   )
 }
-
-
-output "alb_arn" {
-  value = data.external.alb_arn.result.LoadBalancerArn
-}
